fix: changed type of b and h in cuda worldgen data from long to long long to reflect the s64 in jdah's code

+++ src/cuda/cuda-worldgen.h
@@ -38,7 +38,7 @@ extern "C" {
 
     typedef struct CudaWorldgenData {
         float h_b;
-        long h, b;
+        long long h, b;
     } CUDA_WORLDGEN_DATA;
 
     /*
